Clear remaining ruff lints in shared_utils/
- bar_plotter.py: drop dead xtitle/ytitle assignments (4 sites — lint
  F841/RUF059); dict() call -> {} literal (C408); add noqa to the
  second_category_title param that's accepted for API uniformity but
  never read (ARG001, same rationale as tilemap_plotter).
- hover_formatting.py: int(math.floor(x)) -> math.floor(x) (RUF046 —
  math.floor already returns int).
- histogram_utils.py: wrap a 127-char line (E501); dict comprehension ->
  dict.fromkeys (C420).
- s3_manager.py: @lru_cache(maxsize=None) -> @functools.cache (UP033);
  hashlib.md5(usedforsecurity=False) since we're using it as an integrity
  checksum against S3 ETag, not for security (S324); noqa on the
  module-level _s3_client singleton global (PLW0603).
- timing.py: mutable class attrs annotated with ClassVar (RUF012 — _stats
  and _pending_events); noqa on the long-lived trace-file open (SIM115 —
  closed by stop_trace).
- tilemap_plotter.py: drop 3 unused signature params (orientation,
  label_formatter, categories — no callers pass them); noqa the remaining
  2 (first_category_title, second_category_title — callers DO pass
  meaningful values but the body silently hardcodes empty strings;
  a known design debt, preserving current behavior).
- caching.py: noqa the 'shelve file may not exist in read-only mode'
  try/except/pass (S110) — the comment already explains the intent.

+++ postprocessing/resstockpostproc/shared_utils/generic_plotters/hover_formatting.py
@@ -72,7 +72,7 @@ def format_count_value(value: float | int | None) -> str:
     """Format a count-like value as a rounded integer with separators."""
     if value is None:
         return ""
-    rounded = int(math.floor(float(value) + 0.5))
+    rounded = math.floor(float(value) + 0.5)
     return f"{rounded:,}"
